Add scale_text() modifier to get the right size text as a 3D object
Add scale_text(), a helper module that scales and extrudes a text object
to match the scale of the dimension lines, and to extrude it into a 3D
object. This replaces the manual call to scale() that is open coded
everywhere dimlines uses text(), which makes it possible for users of
dimlines to get scaled text the same size as used by dimlines.

Extruding to a 3D object is important so that dimlines output isn't
mixed between 2D and 3D objects. OpenSCAD cannot render or export mixed
2D/3D diagrams, so all objects need to be the same type before
rendering.

Also change DIM_FONTSCALE to DIM_FONTSIZE. The value of DIM_FONTSIZE is
in the range used by the size argument to text(), and is the approximate
height of a single line of text.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>