Use text() halign and valign to position text
The text() module has built in support for positioning text with
left/right/center justification, and top/bottom/center/baseline
alignment. Simplify the text positioning code by using halign and valign
when placing text instead of calculating an approximate size based on
the number of characters. This results in more consistent and accurate
placement of text.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>