Add standard license and copyright notification to header block
For the convenience of anyone using this modules, add license and
copyright notification to the header block so that when this file gets
copied into other projects the important license information remains
with it. The SPDX-style license tag line is used so that the full MIT
license block doesn't need to be cut & pasted into the file.

At the same time tidy up some revision information from the header block
(that's what git is for after all) and add a few notes on how to use the
program.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>